# [5.2] Speed up chunk for large sets of data