--- a/src/Illuminate/Database/Query/Builder.php
+++ b/src/Illuminate/Database/Query/Builder.php
@@ -1323,7 +1323,7 @@ class Builder
      * @param int $lastId
      * @return Builder|static
      */
-    public function forPageById($perPage = 15, $idField, $lastId = 0)
+    public function pageAfterId($perPage = 15, $idField = 'id', $lastId = 0)
     {
         return $this->select($idField)
             ->where($idField, '>', $lastId)
@@ -1642,7 +1642,7 @@ class Builder
     public function chunkById($count, callable $callback, $idField)
     {
         $lastId = null;
-        $results = $this->forPageById($count, $idField)->get();
+        $results = $this->pageAfterId($count, $idField)->get();
 
         while (! $results->isEmpty()) {
             if (call_user_func($callback, $results) === false) {
@@ -1653,7 +1653,7 @@ class Builder
                 $lastId = last($results->all())->{$idField};
             }
 
-            $results = $this->forPageById($count, $idField, $lastId)->get();
+            $results = $this->pageAfterId($count, $idField, $lastId)->get();
         }
 
         return true;