I've been working on a project that involves sifting through some pretty large databases, and the query builder's chunk method was not really doing it for me because it slowed down so much as it worked through the table. 

As chunk gets deeper and deeper into a table, it asks MySQL to count through more and more rows. When chunk says to MySQL, `select whatever from wherever limit 800000, 1000` MySQL counts through 800,000 rows without using an index or nuthin. Then, the next page, it again counts through all previous rows to find the next set. Poor MySQL... It gets really really slow (details below).

I realize this may be too much an edge case, so didn't want to spend too much time working on the code; just looking for feedback at this point.

If you're still reading, here are some scrappy benchmarks. :)

--- 

After 4 seconds, chunk is very optimistic:
`187500/862443 [▓▓▓▓▓▓░░░░░░░░░░░░░░░░░░░░░░]  21% 4 secs/18 secs 10.0 MiB`

However, in the end the process has taken substantially longer:
`862443/862443 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100% 2 mins/2 mins 10.0 MiB`

As you can imagine, the problem gets progressively worse as the database gets larger.

If we instead query by `'id' > $theLastIdFromTheLastSet`, the chunking is much faster.

Again, after 4 seconds:
`387000/862443 [▓▓▓▓▓▓▓▓▓▓▓▓░░░░░░░░░░░░░░░░]  44% 4 secs/9 secs 10.0 MiB`

And the same set takes only 21s in the end.
`862443/862443 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100% 21 secs/21 secs 10.0 MiB`