@GrahamCampbell 2016-03-25
This only works if we are actually ordering by id.