@GrahamCampbell 2016-03-25
Also, the each function is still forced to use the poor performance chunk method.