@taylorotwell 2016-03-25
Yeah it's common enough to order by ID that it's probably worth the improvement here. Thanks!