@kevindoole 2016-03-27
Ok, great!

So far i've kept the changed functionality completely separated in a `chunkById` method. Is it worth spending a bit extra time to make the chunk method use `where` instead of `offset` by default when it's possible, or would you rather keep it separated? I guess the logic would be something like, if there's already an `order by` or if there's no id available, use `offset`. Otherwise, use the `where id > x`. 

Come to think of it, i guess `where id < x order by id desc` is more natural.

Sound good?