@taylorotwell 2016-03-28
I would just keep them separate how you have them now.