@vlakoff 2016-03-29
* Looks like there is code duplication. If you define `chunkById` only in Builder, it will also be callable by Eloquent because of the __call().
* Because of the `select` statement in `pageAfterId`, the results contain only id column...