@halaei 2016-04-12
Could there be a similar thing for pagination?